+++ nix/hermes.nix
@@ -7,17 +7,17 @@
   symlinkJoin,
   openssl,
   pkg-config,
+  apple-sdk_15,
 }:
 rustPlatform.buildRustPackage rec {
   name = "hermes";
   inherit src;
   cargoBuildFlags = "-p ibc-relayer-cli";
   buildInputs = lib.optionals stdenv.isDarwin [
-    darwin.apple_sdk.frameworks.Security
+    apple-sdk_15
     pkg-config
     openssl
     darwin.libiconv
-    darwin.apple_sdk.frameworks.SystemConfiguration
   ];
   cargoLock = {
     lockFile = "${src}/Cargo.lock";